+++ src/common/hr/Hr.jsx
@@ -19,7 +19,7 @@ const StHr = styled.hr`
 				`;
 			case "feedHr":
 				return css`
-					width: 375px;
+					width: 430px;
 					height: 2px;
 					background-color: #efefef;
 				`;
