+++ src/redux/config/configStore.js
@@ -3,7 +3,6 @@ import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
 import join from "../modules/join/joinSlice";
 import todoListSlice from "../modules/todoList/todoListSlice";
 import profileSlice from "../modules/profileSlice";
-import commentSlice from "../modules/commentSlice";
 
 const store = configureStore({
 	reducer: {
@@ -11,7 +10,6 @@ const store = configureStore({
 		todoListSlice,
 		feed,
 		profileSlice,
-		commentSlice,
 	},
 	middleware: getDefaultMiddleware({
 		serializableCheck: false,
