+++ src/redux/modules/feed/feedSlice.js
@@ -115,6 +115,85 @@ export const __uploadFeed = createAsyncThunk(
 	},
 );
 
+// 코멘트 추가
+export const __addComment = createAsyncThunk(
+	"comment/addComment",
+	async (payload, thunkAPI) => {
+		try {
+			await axios.post(
+				`${serverUrl}/api/feed/${payload.id}/comment`,
+				{
+					commentContent: payload.content,
+				},
+				{
+					headers: {
+						Authorization: accessToken,
+					},
+				},
+				{
+					withCredentials: true,
+				},
+			);
+			window.location.replace(`/feed/${payload.feedId}`);
+			return thunkAPI.fulfillWithValue(payload);
+		} catch (error) {
+			return thunkAPI.rejectWithValue(error);
+		}
+	},
+);
+
+// 코멘트 수정
+export const __editComment = createAsyncThunk(
+	"comment/editComment",
+	async (payload, thunkAPI) => {
+		try {
+			const response = await axios.put(
+				`${serverUrl}/api/feed/${payload.feedId}/comment?comment-id=${payload.commentId}`,
+				{
+					headers: {
+						Authorization: accessToken,
+					},
+				},
+				{
+					withCredentials: true,
+				},
+			);
+			alert("수정 완료");
+			window.history.back();
+			return thunkAPI.fulfillWithValue(response);
+		} catch (error) {
+			alert("수정 실패");
+			return thunkAPI.rejectWithValue(error);
+		}
+	},
+);
+
+// 코멘트 삭제
+export const __deleteComment = createAsyncThunk(
+	"comment/deleteComment",
+	async (payload, thunkAPI) => {
+		try {
+			await axios.delete(
+				`${serverUrl}/api/feed/${payload.feedId}/comment?comment-id=${payload.commentId}`,
+				{
+					headers: {
+						Authorization: accessToken,
+					},
+				},
+				{
+					withCredentials: true,
+				},
+			);
+			alert("삭제 완료");
+			window.location.replace(`/feed/${payload.feedId}`);
+			return thunkAPI.fulfillWithValue(payload);
+		} catch (error) {
+			alert("삭제 실패");
+			return thunkAPI.rejectWithValue(error);
+		}
+	},
+);
+
 const initialState = {
 	feedList: [],
 	checkedList: [],
@@ -123,6 +202,7 @@ const initialState = {
 	formPhotoList: [],
 	successTodo: [],
 	feedItem: {},
+	commentList: [],
 };
 
 export const feedSlice = createSlice({
@@ -197,6 +277,16 @@ export const feedSlice = createSlice({
 			// 피드 단건 조회 성공
 			.addCase(__getFeedItem.fulfilled, (state, action) => {
 				state.feedItem = action.payload;
+			})
+			.addCase(__addComment.fulfilled, (state, action) => {
+				state.feedItem.commentResponseDtoList = action.payload;
+				state.feedItem.countComment = action.payload;
+				state.commentList?.push(action.payload);
+			})
+			.addCase(__deleteComment.fulfilled, (state, action) => {
+				state.commentList = state.commentList?.filter(item => {
+					return item !== action.payload;
+				});
 			});
 	},
 });
